requirements.txt
curl-cffi
python-dotenv
rich